Backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# LangChain ecosystem (compatible versions)
langchain>=0.2.0,<0.3.0
langchain-core>=0.2.0,<0.3.0
langchain-community>=0.2.0,<0.3.0
langchain-openai>=0.1.0,<0.2.0
langchain-qdrant>=0.1.0,<0.2.0
langchain-classic>=0.1.0
langgraph>=0.1.0,<0.2.0

# Vector Database
qdrant-client>=1.9.0,<2.0.0

# HTTP Client
requests>=2.31.0

# Utilities
pydantic>=2.5.0,<3.0.0
python-dotenv==1.0.0
typing_extensions>=4.0.0
